requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple
--trusted-host mirrors.aliyun.com

aliyun-python-sdk-core-v3==2.9.4  # via aliyun-python-sdk-kms, oss2
aliyun-python-sdk-kms==2.5.0  # via oss2
arrow==0.12.1
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
crcmod==1.7               # via oss2
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.3
django-rest-swagger==2.2.0
django==2.1.2
djangorestframework==3.8.2
factory-boy==2.11.1
faker==0.9.1              # via factory-boy
idna==2.7                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
markupsafe==1.0           # via jinja2
mysqlclient==1.3.13
openapi-codec==1.3.2      # via django-rest-swagger
oss2==2.5.0
pip-tools==3.1.0
pycryptodome==3.6.6       # via aliyun-python-sdk-core-v3, oss2
python-dateutil==2.7.3    # via arrow, faker
pytz==2018.5              # via django
raven==6.9.0
requests==2.19.1          # via coreapi, oss2
simplejson==3.16.0        # via django-rest-swagger
six==1.11.0               # via django-extensions, faker, pip-tools, python-dateutil
text-unidecode==1.2       # via faker
uritemplate==3.0.0        # via coreapi
urllib3==1.23             # via requests
werkzeug==0.14.1
